refactor: 딜러와 플레이어의 중복 로직 제거 및 패키지 분리
