+++ src/main/java/domain/gamer/Players.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.gamer;
 
 import java.util.ArrayList;
 import java.util.List;
