+++ src/main/java/view/InputView.java
@@ -3,7 +3,7 @@ package view;
 import java.util.List;
 import java.util.Scanner;
 
-import domain.Player;
+import domain.gamer.Player;
 
 public class InputView {
 	private final Scanner scanner;
