+++ src/test/java/domain/CardHandTest.java
@@ -7,6 +7,11 @@ import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
+import domain.card.Card;
+import domain.card.CardHand;
+import domain.card.Rank;
+import domain.card.Suit;
+
 public class CardHandTest {
 
 	@DisplayName("카드 패의 합을 구한다.")
