+++ src/test/java/domain/gamer/PlayerTest.java
@@ -1,4 +1,4 @@
-package domain;
+package domain.gamer;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatCode;
@@ -11,6 +11,10 @@ import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.ValueSource;
+
+import domain.card.Card;
+import domain.card.Rank;
+import domain.card.Suit;
 
 public class PlayerTest {
 
@@ -67,6 +71,6 @@ public class PlayerTest {
 		player.receiveCard(new Card(Suit.HEART, Rank.Q));
 
 		// when & then
-		assertThat(player.isBurst()).isTrue();
+		assertThat(player.isBust()).isTrue();
 	}
 }
